Removed babel dependency
