--- .babelrc
@@ -1,10 +1,0 @@
-{
-  "presets": ["env"],
-  "plugins": [
-//    "transform-decorators",
-    "transform-decorators-legacy",
-    ["babel-plugin-transform-builtin-extend", {
-      "globals": ["Error", "Array"]
-    }]
-  ]
-}
