+++ core/errors/index.js
@@ -1,0 +1,9 @@
+const InvalidArgumentError = require('./invalid-argument-error')
+const NonexistentEntityError = require('./nonexistent-entity-error')
+const NotImplementedError = require('./not-implemented-error')
+
+module.exports = {
+  InvalidArgumentError,
+  NonexistentEntityError,
+  NotImplementedError
+}
