+++ core/errors/invalid-argument-error.js
@@ -1,0 +1,5 @@
+const AbstractError = require('../../abstract-error')
+
+class InvalidArgumentError extends AbstractError {}
+
+module.exports = InvalidArgumentError
