+++ core/errors/nonexistent-entity-error.js
@@ -1,0 +1,5 @@
+const AbstractError = require('../../abstract-error')
+
+class NonexistentEntityError extends AbstractError {}
+
+module.exports = NonexistentEntityError
