+++ core/errors/not-implemented-error.js
@@ -1,0 +1,5 @@
+const AbstractError = require('../../abstract-error')
+
+class NotImplementedError extends AbstractError {}
+
+module.exports = NotImplementedError
