+++ core/index.js
@@ -1,0 +1,9 @@
+const Holder = require('./holder')
+const Repository = require('./repository')
+const User = require('./user')
+
+module.exports = {
+  Repository,
+  Holder,
+  User
+}
