+++ db/index.js
@@ -1,0 +1,5 @@
+const Repository = require('./sequelize-repository')
+
+module.exports = {
+  Repository
+}
