+++ server/errors/index.js
@@ -1,0 +1,5 @@
+const HeaderValidationError = require('./header-validation-error')
+
+module.exports = {
+  HeaderValidationError
+}
