+++ server/handlers/index.js
@@ -1,0 +1,7 @@
+const ErrorHandler = require('./error-handler')
+const UserHandler = require('./user-handler')
+
+module.exports = {
+  ErrorHandler,
+  UserHandler
+}
