+++ server/routers/index.js
@@ -1,0 +1,5 @@
+const userRouter = require('./user-router')
+
+module.exports = {
+  userRouter
+}
