--- src/core/errors/invalid-argument-error.js
@@ -1,3 +1,0 @@
-import AbstractError from '../../abstract-error'
-
-export default class InvalidArgumentError extends AbstractError {}
