--- src/core/errors/nonexistent-entity-error.js
@@ -1,3 +1,0 @@
-import AbstractError from '../../abstract-error'
-
-export default class NonexistentEntityError extends AbstractError {}
