--- src/core/errors/not-implemented-error.js
@@ -1,3 +1,0 @@
-import AbstractError from '../../abstract-error'
-
-export default class NotImplementedError extends AbstractError {}
