--- src/core/index.js
@@ -1,4 +1,0 @@
-export { default as errors } from './errors'
-export { default as Repository } from './repository'
-export { default as Holder } from './holder'
-export { default as User } from './user'
