--- src/db/index.js
@@ -1,1 +1,0 @@
-export { default as Repository } from './sequelize-repository'
