--- src/server/errors/header-validation-error.js
@@ -1,9 +1,0 @@
-import AbstractError from '../../abstract-error'
-
-export default class HeaderValidationError extends AbstractError {
-  constructor (header, message, cause) {
-    super(message, cause)
-    this.header = header
-    this.errors = [{header: header, message: message}]
-  }
-}
