--- src/server/errors/index.js
@@ -1,1 +1,0 @@
-export { default as HeaderValidationError } from './header-validation-error'
