--- src/server/handlers/index.js
@@ -1,2 +1,0 @@
-export { default as ErrorHandler } from './error-handler'
-export { default as UserHandler } from './user-handler'
