--- src/server/routers/index.js
@@ -1,1 +1,0 @@
-export { default as userRouter } from './user-router'
